requirements.txt

jmespath==1.0.1

setuptools==68.1.2

redis>=4.0.0

# Core Web Framework & ORM
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.10  # PostgreSQL driver
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
uvicorn[standard]

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-graphql-jwt==0.4.0  # Optional if using GraphQL JWT

# GraphQL Support
graphene==3.3
graphene-django==3.1.6
graphql-core==3.2.3
graphql-relay==3.2.0

# Task Queues & Scheduling
django-celery-beat==2.6.0
schedule==1.2.0

# AWS & Storage
boto3==1.38.36
botocore==1.38.36
s3transfer==0.13.0
watchtower==3.0.1
django-storages==1.14

# PDF & File Handling
openpyxl==3.1.2
XlsxWriter==3.1.9
pytesseract==0.3.10
opencv-python==4.9.0.80
pdfkit==1.0.0
pydyf>=0.11.0
pychromepdf==1.1
xlrd==2.0.1

# Utilities & Others
Pillow==10.4.0
python-dotenv==1.0.1
Markdown==3.6
numpy==2.3.0
num2words==0.5.13
fiscalyear==0.4.0
jsonfield==3.1.0
random-password-generator==2.2.0
easygui==0.98.3
tk==0.1.0

# HTTP/Networking
requests==2.31.0
certifi==2024.2.2
idna==3.7
charset-normalizer==3.3.2
urllib3==2.4.0

# PDF/Text extraction
textract-trp==0.1.3
selenium==4.21.0
webdriver-manager==4.0.2

# Development & Deployment
gunicorn==21.2.0
whitenoise==6.6.0
sqlparse==0.5.0
asgiref==3.8.1
six==1.17.0
python-dateutil==2.9.0.post0
PyJWT==2.8.0

cryptography==45.0.4
razorpay==1.4.2
pandas==2.3.0
jinja2==3.1.6
weasyprint==66.0